Move rancher-ui to components
